requirements.txt
langchain>=0.1.0
langchain-groq
langchain-community
langgraph
python-dotenv
pydantic
flask
uvicorn